# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * EmberRealm - Anime & Manga streaming platform API proxy
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AnimeItem {
  id: number;
  title: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  posterPath?: string | null;
  /** @nullable */
  backdropPath?: string | null;
  /** @nullable */
  overview?: string | null;
  /** @nullable */
  voteAverage?: number | null;
  /** @nullable */
  voteCount?: number | null;
  /** @nullable */
  firstAirDate?: string | null;
  /** @nullable */
  releaseDate?: string | null;
  /** @nullable */
  mediaType?: string | null;
  genreIds?: number[];
}

export interface AnimeListResponse {
  results: AnimeItem[];
  page: number;
  totalPages: number;
  totalResults: number;
}

export interface Genre {
  id: number;
  name: string;
}

export interface Season {
  id: number;
  name: string;
  seasonNumber: number;
  /** @nullable */
  episodeCount?: number | null;
  /** @nullable */
  airDate?: string | null;
  /** @nullable */
  overview?: string | null;
  /** @nullable */
  posterPath?: string | null;
}

export interface CastMember {
  id: number;
  name: string;
  /** @nullable */
  character?: string | null;
  /** @nullable */
  profilePath?: string | null;
}

export type AnimeDetailsNetworksItem = {
  id: number;
  name: string;
};

export interface AnimeDetails {
  id: number;
  name: string;
  /** @nullable */
  originalName?: string | null;
  /** @nullable */
  posterPath?: string | null;
  /** @nullable */
  backdropPath?: string | null;
  /** @nullable */
  overview?: string | null;
  /** @nullable */
  voteAverage?: number | null;
  /** @nullable */
  voteCount?: number | null;
  /** @nullable */
  firstAirDate?: string | null;
  /** @nullable */
  lastAirDate?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  numberOfSeasons?: number | null;
  /** @nullable */
  numberOfEpisodes?: number | null;
  genres?: Genre[];
  seasons?: Season[];
  cast?: CastMember[];
  networks?: AnimeDetailsNetworksItem[];
  recommendations?: AnimeItem[];
}

export interface VideoItem {
  id: string;
  name: string;
  key: string;
  site: string;
  type: string;
  official: boolean;
}

export interface VideoListResponse {
  results: VideoItem[];
}

export interface Episode {
  id: number;
  name: string;
  episodeNumber: number;
  seasonNumber: number;
  /** @nullable */
  overview?: string | null;
  /** @nullable */
  airDate?: string | null;
  /** @nullable */
  stillPath?: string | null;
  /** @nullable */
  voteAverage?: number | null;
  /** @nullable */
  runtime?: number | null;
}

export interface EpisodeListResponse {
  episodes: Episode[];
  seasonNumber: number;
  name: string;
}

export interface EmbedInfo {
  embedUrl: string;
  type: string;
  tmdbId: number;
  episode: number;
  season: number;
}

export interface GenreListResponse {
  genres: Genre[];
}

export interface MangaItem {
  id: string;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  coverUrl?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  year?: number | null;
  tags?: string[];
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  lastChapter?: string | null;
  /** @nullable */
  lastUpdated?: string | null;
}

export interface MangaListResponse {
  results: MangaItem[];
  total: number;
  limit: number;
  offset: number;
}

export interface MangaDetails {
  id: string;
  title: string;
  altTitles?: string[];
  /** @nullable */
  description?: string | null;
  /** @nullable */
  coverUrl?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  year?: number | null;
  tags?: string[];
  authors?: string[];
  artists?: string[];
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  totalChapters?: number | null;
  /** @nullable */
  lastUpdated?: string | null;
}

export interface ChapterItem {
  id: string;
  /** @nullable */
  chapter?: string | null;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  volume?: string | null;
  /** @nullable */
  publishAt?: string | null;
  /** @nullable */
  readableAt?: string | null;
  /** @nullable */
  pages?: number | null;
  translatedLanguage: string;
  /** @nullable */
  scanlationGroup?: string | null;
}

export interface ChapterListResponse {
  chapters: ChapterItem[];
  total: number;
  limit: number;
  offset: number;
}

export interface ChapterPagesResponse {
  chapterId: string;
  baseUrl: string;
  hash: string;
  pages: string[];
  dataSaver: boolean;
}

export type GetTrendingAnimeParams = {
page?: number;
};

export type GetPopularAnimeParams = {
page?: number;
};

export type GetTopRatedAnimeParams = {
page?: number;
};

export type GetSeasonalAnimeParams = {
page?: number;
};

export type SearchAnimeParams = {
q: string;
page?: number;
};

export type SearchMangaParams = {
q: string;
limit?: number;
offset?: number;
};

export type GetPopularMangaParams = {
limit?: number;
offset?: number;
};

export type GetLatestMangaParams = {
limit?: number;
offset?: number;
};