# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EmberRealm - Anime & Manga streaming platform API proxy
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /anime/trending:
    get:
      operationId: getTrendingAnime
      tags: [anime]
      summary: Get trending anime from TMDB
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Trending anime list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnimeListResponse"
  /anime/popular:
    get:
      operationId: getPopularAnime
      tags: [anime]
      summary: Get popular anime from TMDB
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Popular anime list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnimeListResponse"
  /anime/top-rated:
    get:
      operationId: getTopRatedAnime
      tags: [anime]
      summary: Get top rated anime from TMDB
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Top rated anime list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnimeListResponse"
  /anime/seasonal:
    get:
      operationId: getSeasonalAnime
      tags: [anime]
      summary: Get currently airing anime from TMDB
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Seasonal anime list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnimeListResponse"
  /anime/search:
    get:
      operationId: searchAnime
      tags: [anime]
      summary: Search anime by query
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnimeListResponse"
  /anime/{id}:
    get:
      operationId: getAnimeDetails
      tags: [anime]
      summary: Get detailed anime info from TMDB
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Anime details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnimeDetails"
  /anime/{id}/recommendations:
    get:
      operationId: getAnimeRecommendations
      tags: [anime]
      summary: Get anime recommendations from TMDB
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recommendations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnimeListResponse"
  /anime/{id}/videos:
    get:
      operationId: getAnimeVideos
      tags: [anime]
      summary: Get trailers and videos from TMDB
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Videos list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoListResponse"
  /anime/{id}/seasons/{seasonNumber}/episodes:
    get:
      operationId: getAnimeEpisodes
      tags: [anime]
      summary: Get episode list for a season from TMDB
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: path
          name: seasonNumber
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Episode list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EpisodeListResponse"
  /anime/{id}/embed/{season}/{episode}:
    get:
      operationId: getAnimeEmbedUrl
      tags: [anime]
      summary: Get embed URL for anime episode playback
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: path
          name: season
          required: true
          schema:
            type: integer
        - in: path
          name: episode
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Embed URL info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmbedInfo"
...
components:
  schemas:
...
    AnimeItem:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        name:
          type: ["string", "null"]
        posterPath:
          type: ["string", "null"]
        backdropPath:
          type: ["string", "null"]
        overview:
          type: ["string", "null"]
        voteAverage:
          type: ["number", "null"]
        voteCount:
          type: ["integer", "null"]
        firstAirDate:
          type: ["string", "null"]
        releaseDate:
          type: ["string", "null"]
        mediaType:
          type: ["string", "null"]
        genreIds:
          type: array
          items:
            type: integer
      required:
        - id
        - title
    AnimeListResponse:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/AnimeItem"
        page:
          type: integer
        totalPages:
          type: integer
        totalResults:
          type: integer
      required:
        - results
        - page
        - totalPages
        - totalResults
    Genre:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
      required:
        - id
        - name
    Season:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        seasonNumber:
          type: integer
        episodeCount:
          type: ["integer", "null"]
        airDate:
          type: ["string", "null"]
        overview:
          type: ["string", "null"]
        posterPath:
          type: ["string", "null"]
      required:
        - id
        - name
        - seasonNumber
    CastMember:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        character:
          type: ["string", "null"]
        profilePath:
          type: ["string", "null"]
      required:
        - id
        - name
    AnimeDetails:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        originalName:
          type: ["string", "null"]
        posterPath:
          type: ["string", "null"]
        backdropPath:
          type: ["string", "null"]
        overview:
          type: ["string", "null"]
        voteAverage:
          type: ["number", "null"]
        voteCount:
          type: ["integer", "null"]
        firstAirDate:
          type: ["string", "null"]
        lastAirDate:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        numberOfSeasons:
          type: ["integer", "null"]
        numberOfEpisodes:
          type: ["integer", "null"]
        genres:
          type: array
          items:
            $ref: "#/components/schemas/Genre"
        seasons:
          type: array
          items:
            $ref: "#/components/schemas/Season"
        cast:
          type: array
          items:
            $ref: "#/components/schemas/CastMember"
        networks:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              name:
                type: string
            required:
              - id
              - name
        recommendations:
          type: array
          items:
            $ref: "#/components/schemas/AnimeItem"
      required:
        - id
        - name
    VideoItem:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        key:
          type: string
        site:
          type: string
        type:
          type: string
        official:
          type: boolean
      required:
        - id
        - name
        - key
        - site
        - type
        - official
    VideoListResponse:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/VideoItem"
      required:
        - results
    Episode:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        episodeNumber:
          type: integer
        seasonNumber:
          type: integer
        overview:
          type: ["string", "null"]
        airDate:
          type: ["string", "null"]
        stillPath:
          type: ["string", "null"]
        voteAverage:
          type: ["number", "null"]
        runtime:
          type: ["integer", "null"]
      required:
        - id
        - name
        - episodeNumber
        - seasonNumber
    EpisodeListResponse:
      type: object
      properties:
        episodes:
          type: array
          items:
            $ref: "#/components/schemas/Episode"
        seasonNumber:
          type: integer
        name:
          type: string
      required:
        - episodes
        - seasonNumber
        - name
    EmbedInfo:
      type: object
      properties:
        embedUrl:
          type: string
        type:
          type: string
        tmdbId:
          type: integer
        episode:
          type: integer
        season:
          type: integer
      required:
        - embedUrl
        - type
        - tmdbId
        - episode
        - season
...