# lib/api-zod/src/generated/types/animeListResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * EmberRealm - Anime & Manga streaming platform API proxy
 * OpenAPI spec version: 0.1.0
 */
import type { AnimeItem } from './animeItem';

export interface AnimeListResponse {
  results: AnimeItem[];
  page: number;
  totalPages: number;
  totalResults: number;
}