# lib/api-zod/src/generated/types/embedInfo.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * EmberRealm - Anime & Manga streaming platform API proxy
 * OpenAPI spec version: 0.1.0
 */

export interface EmbedInfo {
  embedUrl: string;
  type: string;
  tmdbId: number;
  episode: number;
  season: number;
}