# lib/api-zod/src/generated/types/episode.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * EmberRealm - Anime & Manga streaming platform API proxy
 * OpenAPI spec version: 0.1.0
 */

export interface Episode {
  id: number;
  name: string;
  episodeNumber: number;
  seasonNumber: number;
  /** @nullable */
  overview?: string | null;
  /** @nullable */
  airDate?: string | null;
  /** @nullable */
  stillPath?: string | null;
  /** @nullable */
  voteAverage?: number | null;
  /** @nullable */
  runtime?: number | null;
}