# lib/api-zod/src/generated/types/genre.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * EmberRealm - Anime & Manga streaming platform API proxy
 * OpenAPI spec version: 0.1.0
 */

export interface Genre {
  id: number;
  name: string;
}