# lib/api-zod/src/generated/types/videoItem.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * EmberRealm - Anime & Manga streaming platform API proxy
 * OpenAPI spec version: 0.1.0
 */

export interface VideoItem {
  id: string;
  name: string;
  key: string;
  site: string;
  type: string;
  official: boolean;
}