# lib/api-zod/src/generated/types/videoListResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * EmberRealm - Anime & Manga streaming platform API proxy
 * OpenAPI spec version: 0.1.0
 */
import type { VideoItem } from './videoItem';

export interface VideoListResponse {
  results: VideoItem[];
}